Add files via upload
In this document we build the general Fibonacci word Sn shown in the Wikipedia article and its cutting sequence (see more about this in https://en.wikipedia.org/wiki/Fibonacci_word). Thereafter we will use the same algorithm for the construction of cutting sequences for the binary sequence from written texts (by some suitable association rule) in order to obtain a characterization as a function of the equation of the cutting sequence.