Add files via upload
This work aims to characterize written texts to verify the distinction between natural language and synthetically generated language. It seeks to establish a method to discern texts that obey grammatical rules from those randomly generated texts. To verify this we are going to compare the text with the shuffled case of the book.